Add remove method to title and text
In the 'savePost' and 'savePage' method I added the remove method to
both so that the items will be quickly removed but then reappear
quickly due to the getContents api. It happens so quickly that it's not
noticeable. I did this because it wasn't editing in real time
consistently and found that this method works every time.